# app/routes/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Swagger Policy Street Demo",
    "description": "User management system",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3500/api/v1"
    }
  ],
...
  "paths": {
...
    "/users": {
      "post": {
        "tags": ["users"],
        "summary": "Create a new user",
        "operationId": "createUser",
        "requestBody": {
          "description": "User object that needed to add to the system",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Any of the parameters is missing",
            "content": {}
          },
          "500": {
            "description": "Unexpected error",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "required": ["login", "password", "user_type"],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "user_type": {
            "type": "string",
            "description": "pet status in the store",
            "enum": ["superadmin", "admin", "user"]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "x-auth-token",
        "in": "header"
      }
    }
  }
}